Restro_admin/requirements.txt
django
djangorestframework
grpcio
grpcio-tools
djangorestframework-simplejwt
python-dotenv
psycopg